pyproject.toml
[project]
name = "api-queue"
version = "0.1.0"
description = "High-throughput attendee scoring API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.5",
    "djangorestframework>=3.16.1",
    "gunicorn>=23.0.0",
    "joblib>=1.5.1",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "psycopg[binary]>=3.2.9",
    "scikit-learn>=1.7.1",
    "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-django>=4.11.1",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
